feat: resource drop system and user feedback during collection
Added a resource drop system to the collection skill. Now, when the player collects trash, one of four materials can be generated:

Plastic (~43%)

Metal (~26%)

Electronics (~13%)

Mechanical parts (~17%)

These resources will be used in a future update to upgrade the player's skills.

Also added visual feedback to inform the player which material was collected.